Move PC+ShowRoute tests to own file
