+++ Sources/App/Controllers/PackageController+routes.swift
@@ -34,6 +34,7 @@ struct PackageController {
             }
     }
 
+    #warning("add test")
     func readme(req: Request) throws -> EventLoopFuture<Node<HTML.BodyContext>> {
         guard
             let owner = req.parameters.get("owner"),
@@ -54,6 +55,7 @@ struct PackageController {
             .map { PackageReadme.View(model: $0).document() }
     }
     
+#warning("add test")
     func releases(req: Request) throws -> EventLoopFuture<Node<HTML.BodyContext>> {
         guard
             let owner = req.parameters.get("owner"),
